a lot of work have been done

+++ CinemaApp/Controllers/CinemaController.cs
@@ -32,17 +32,75 @@ namespace CinemaApp.Controllers
             return new cinemaDto(cinema);
         }
         [HttpPost]
-        public async Task CreateCinemaAsync(cinemaDto cinemaDto)
+        public async Task<IActionResult> CreateCinemaAsync(CreateCinemaRequest request)
         {
-            var cinema = new Cinema { Id=Guid.NewGuid(), Name = cinemaDto.Name, Address = cinemaDto.Address};
+            string? imagePath = null;
+            var cinema = new Cinema
+            {
+                Id = Guid.NewGuid(),
+                Name = request.CinemaDto.Name,
+                Address = request.CinemaDto.Address,
+                ImagePath = imagePath
+            };
+            int numOfHalls = Convert.ToInt32(request.NumOfHalls);
+            Console.WriteLine(request.NumOfHalls);
             await _context.Cinemas.AddAsync(cinema);
+
+            for (int i = 0; i < numOfHalls; i++)
+            {
+                var hall = new Hall
+                {
+                    Id = Guid.NewGuid(),
+                    CinemaId = cinema.Id,
+                    Number = (i + 1).ToString(),
+                    Seats = 100,
+                    Cinema = cinema,
+                    Sessions = new List<Session>()
+                };
+                await _context.Halls.AddAsync(hall);
+            }
+
             await _context.SaveChangesAsync();
+            return Ok(cinema);
+        }
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UploadImageAsync(Guid id, IFormFile? formFile)
+        {
+            var cinema = await _context.Cinemas.FindAsync(id);
+            string? imagePath = null;
+
+            if (formFile != null)
+            {
+                var fileName = $"{Guid.NewGuid()}_{formFile.FileName}";
+                var directoryPath = Path.Combine("client-app", "public", "images");
+                if (!Directory.Exists(directoryPath))
+                {
+                    Directory.CreateDirectory(directoryPath);
+                }
+
+                var filePath = Path.Combine(directoryPath, fileName);
+
+                using (var stream = new FileStream(filePath, FileMode.Create))
+                {
+                    await formFile.CopyToAsync(stream);
+                }
+
+                imagePath = $"/images/{fileName}";
+                cinema.ImagePath = imagePath;
+            }
+
+            await _context.SaveChangesAsync();
+            return Ok(imagePath);
         }
         [HttpPut]
         public async Task UpdateCinemaAsync(cinemaDto cinemaDto)
         {
             var cinema = await _context.Cinemas.FirstOrDefaultAsync(c => c.Id == cinemaDto.Id);
-            cinema.Name=cinemaDto.Name;
+            if (cinema == null)
+            {
+                throw new Exception("Cinema not found");
+            }
+            cinema.Name = cinemaDto.Name;
             cinema.Address = cinemaDto.Address;
             _context.Cinemas.Update(cinema);
             await _context.SaveChangesAsync();
@@ -51,8 +109,23 @@ namespace CinemaApp.Controllers
         public async Task DeleteCinemaAsync(Guid id)
         {
             var cinema = await _context.Cinemas.FirstOrDefaultAsync(c => c.Id == id);
+            if (cinema == null)
+            {
+                throw new Exception("Cinema not found");
+            }
+
+            if (!string.IsNullOrEmpty(cinema.ImagePath))
+            {
+                var filePath = Path.Combine("client-app", "public", "images",cinema.ImagePath.TrimStart('/'));
+                if (System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.Delete(filePath);
+                }
+            }
+
             _context.Cinemas.Remove(cinema);
             await _context.SaveChangesAsync();
+
         }
     }
 }
